Update patch set 1
Patch Set 1: This need some tweaks before it is merged

I would strongly recommend to factor the stuff that is in the script into a makefile and also not to commit generated file. There is a place for generated files in "workdir".

Patch-set: 1
Reviewer: Gerrit User 1000005 <1000005@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0
Label: Code-Review=-1