Update patch set 2
Patch Set 2:

>> 4) I have my doubts about time-zoned time without a date.
>>     How do you convert 20:00+00 to time zone +10?
>>     Just "wrap around" modulo 24, i.e. make it 06:00+10?

> yes

>> If you answer yes to that question (that is, always compute modulo 24),
>> then what is the difference between 06:00+11 and 06:00-13?
>> The difference only makes sense if one speaks of dates...
>> Which says that timezones don't make sense unless we know at what date
>> to apply them.

> there is no difference: both denote the same set of inifinitely many instants.

For the sake of the following paragraphs, I'm going to assume we want to keep
the timezone in the struct (as opposed to converting timezoned values to UTC).

As far as reasonably possible, data structures should have a canonical format:
Don't artificially introduce different values that denote the same. If we really
are to consider that a time is a value in timeline quotiented by 24h interval
(that is a representative of the equivalence class "all moments in the timeline
separated by 24h intervals), then we have no reason to allow (for the timezone
value) anything overlapping. Your code currently says everything has to be
between -840 and +840 inclusive. With this interpretation, we should make
one of the following choices:
 - timezone between -719 and +720
 - timezone between -720 and +719
 - timezone between 0 and 1439

That's the logical conclusion of the interpretation you propose.

What I had in mind was another interpretation (which I think
is consistent with the SQL standard's behaviour):

 - a timezone region is a part of the earth surface, politically
   defined, that has the same time (offset to UTC); that is
   calendars and clocks agree all across this zone).

   (Many specific points on the surface of the globe hop from
    one timezone to another twice a year.)

 - a datetime is a moment in time in some timezone region.

 - a datetime is made of two parts, namely a date and a time.

 - the timezone is attached to the time part of datetime.

In this interpretation, the timezone value between -840 and +840
refers to a region of earth. Given two timezone regions, whether
clocks agree between these two timezones is a valid question,
but not the only one and besides the point for data encoding;
they are different timezone regions. E.g. timezone region +660
(+11 hours) and timezone region -780 (-13 hours). Clocks agree
across these two timezones, but calendars don't, so they are
different timezone regions where clocks just happen to agree.

We could make the following valid choice / observation:
XMLSchema-2 has fucked up semantics on time / date / ...,
but for adherence to standards and interoperability, we
"need" to adhere to it rather than doing something more
correct.

>> Is there any part of ODF/OOXML/any other format where we
>> need to import or export a date with timezone?

>> dunno about OOXML

Then bring (one of) our resident expert(s) in this discussion.

> (...) ODF at least has numerous attributes of type date.
> it says nothing about time zones so by referencing the XMLSchema-2
> types they are allowed _everywhere_ a non-time-zoned date is allowed.

So basically you are saying our hands are tied, we need to cover everything
XMLSchema-2 allows. Taking your word on that, further discussion on the
following points is pointless, we need
 - timezoned date in the range -840 to +840
 - timezoned time in the range -840 to +840
 - timezoned datetime in the range -840 to +840
 - untimezoned date, time and datetime.
 - Base will have to map these concepts to the
   SQL world to the best of possibilities. E.g., ignore
   (drop) the timezone on timezoned dates, or
   throw an exception when given a timezoned date
   (require only untimezoned date).



The remaining decision to be made is whether the UNO structs
should encode the timezone explicitly or be "all timezoned values
are in UTC time". The issue is muddled a bit because the UNO
structs are both our internal working structures and our "external"
(scripting) API date types. I'll discuss the two uses separately.

I'm really torn on this issue. Seeing LibreOffice as a programming
platform, the UNO structs are the data types of our "standard
library". By putting explicit timezone in the structs, we make
the programmer's life easier because (s)he does not have to
deal with converting to UTC before making an API call. We also
make hir life more difficult because (s)he has to deal with non-UTC
time for each *result* of an API call.

Here's what I would consider optimal:

 - All our (date)time API calls take and return UTC (date)time or
   untimezoned (date)time. That is, our (date)time structure
   does not have a timezone in it. It is UTC or untimezoned.

 - We provide a UNO service that provides utility functions that

   1) convert between timezoned (date)time on the one hand
       and untimezoned time in a caller-provided timezone on
       the other hand (*two* incoming arguments: a (date)time
       *and* an offset to UTC).

   2) convert between timezoned (date)time on the one hand
       and untimezoned time in local timezone on the other hand
       (*one* incoming argument: a (date)time).

 - Hmm... As an alternative / complement to the above service,
   the (date)time struct constructors takes an optional argument
   that says "the values I'm giving are UTC / localtime / untimezoned"
   (we add a new enum with these three values), and the constructor
   makes the appropriate localtime->UTC conversion.

   Can we do that in UNO? I'm unsure. Actually, I don't expect so.

The spanner in the works, and the reason I'm torn, is timezoned dates.
They don't fit the above schema. I hate them. For timezoned dates,
we *have* to include an explicit timezone in the struct, or expect that
code will store timezoned dates as a pair of values (a date and an
offset) - bleh. This gives us an inconsistent interface between date
on one hand and (date)time on the other hand.

So, the choice boils down to:

 - inconsistent, but best for (date)time: use UTC for timezoned data.
   explicit timezone for timezoned date.

 - consistent, but unnecessarily annoying for (date)time: add explicit
   timezone everywhere. Provide a UNO service that answers the
   question "is this the same point in time"? (that is, semantic equality)

Do we *really* *really* need timezoned dates, or are they a bug
in the ODF specification?



Now, I turn to our internal use. IMHO, untimezoned values should just
be incomparable to timezoned values. Throw an error if it comes up.
The user of the values should "know" / decide what to do:
 - strip / discard timezone
 - convert to/from localtime
 - convert to/from UTC time
 - throw an error
 - ...

In particular, different file formats may have different requirements.
Different parts of code may have different requirements. Don't force
anything on them. Provide utility functions to do the conversion and
that's it.

Having *two* sets of structs (one for timezoned and one for untimezoned)
has the advantage that nobody will "naively" try to compare them...
IMHO it fits this model better. No need to introduce an exception for
"incomparable date/time/datetime value" and no need to handle
that exception everywhere.

Patch-set: 2
Reviewer: Gerrit User 1000084 <1000084@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0