Update patch set 2
Patch Set 2:

>> Don't artificially introduce different values that denote the same. If we really
>> are to consider that a time is a value in timeline quotiented by 24h interval
>> (that is a representative of the equivalence class "all moments in the timeline
>> separated by 24h intervals), then we have no reason to allow (for the timezone
>> value) anything overlapping. Your code currently says everything has to be
>> between -840 and +840 inclusive. With this interpretation, we should make
>> one of the following choices:
>>  - timezone between -719 and +720
>>  - timezone between -720 and +719
>>  - timezone between 0 and 1439

> That simply doesn't match reality. Currently we have values from UTC-11
> (SST Samoa Standard Time) to UTC+14 (LINT Line Islands Time)
> So at least (!) values between -720 and +840 need to be supported,
> rather -840 to +840.

You removed crucial context, namely, that the discussion was *time* values
(and not date or datetime), *and* that the semantics of time had been defined
as:

>>>> Time is a set of infinitely many points on the "universal" UTC time-line, with 24 hour distance between them.
>>>> a Time with timezone is simply offset that much (modulo 24:00).

You see? An offset modulo 24:00, i.e. an offset modulo 1440 minutes. By this definition, it makes no sense to store offset values in a range bigger than 1440. I repeat: *by* *this* *definition*.

Because the value 04:12:20+13 and the value 04:12:20-11 denote (again: *by* *this* *definition*) the same infinitely many points on the same time-line, it is bad data structure design to code them differently. Obviously, when parsing string, we can be *much* more liberal, and even accept e.g. "04:12:20+253". But when parsing "04:12:20+253" (253 = 10*24 + 13) , we wouldn't put "15180" (253 * 60) in the offset, right? So why allow 780?


For DateTime, *obviously* '2009-05-04 04:12:20+13' and '2009-05-04 04:12:20-13' are not the same.


If we take a *different* definition, for example one of:

 - a time with timezone is the time-part of a datetime with timezone
 - a time with timezone is a time in an unknown date;
   the date can be "filled in" by providing a date without timezone or
   date with the *same* timezone

then *yes*, going from UTC-14 to UTC+14 for times makes sense. But *not* with the definition that was in context.

Patch-set: 2