database suggestname: no '/' for queries
In particular, this keeps the new query wizard
from silently failing by default
when the underlying table name has a slash: it would
generate a name proposal based on the table name,
and then use suggestName to get a similar,
yet guaranteed allowed name.

In this case, the "guaranteed allowed" part was failing,
since suggestName failed to strip slashes, which are
forbidden in query names.

Change-Id: I53d3976979ee45f5659765ed5dd3371b78837b5b
Reviewed-on: https://gerrit.libreoffice.org/4709
Reviewed-by: Fridrich Strba <fridrich@documentfoundation.org>
Tested-by: Fridrich Strba <fridrich@documentfoundation.org>