Update patch set 2
Patch Set 2:

> Heh, I wonder if such table names are supported also by the other functionality, e.g. queries, views. Anyway, this fix allows users to rename them at least ;-)

That's up to the database engine. Essentially, you have to understand that "Query" is an SQL command that is stored by LibreOffice in the .odb file. So the "name" of that query is something that is up to LibreOffice, the database engine never sees it. That's why "SELECT * from qryName" will not work, because this will send qryName to the database engine, and it does not know it.

As far as queries are concerned, code archeology shows that this special treatment of '/' in query names dates back from a "new feature: hierarchical queries". I have no clue what this feature is :)

However, views are a database engine feature. They are an SQL command that is stored by the database engine in the database. Their name is up to the database engine. As such, they are just like tables to LibreOffice: LibreOffice selects, updates, deletes, etc from them directly (as far as supported by the database engine).

Patch-set: 2