Update patch set 2
Patch Set 2: Do not merge

So, on IRC the discussion restarted, took other people in, and they want yet other designs.

Given the tight time deadline and the wildly varying opinions, after erack threw in the towel and left IRC, it was more or less consensus-decided to postpone this for 4.2, and have a non-hurried discussion where other people have the time to come in.

Patch-set: 2
Label: Code-Review=-2