Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

OK, this looks great. Lot of work required for that seemingly small tabpage. Just two small things.

a) Keep the %VARIABLE stuff in the new .ui stuff. There was a bug in the %PRODUCTNAME replacement for items in lists which I fixed now, so that probably was the problem there. So I'll push a follow up patch to restore those variables again.

b) Use "bold" not "semibold" for the frame titles, cause that's what we used for all the other ones.

c) The hungarian (light version) variable naming scheme in LibreOffice is aFoo for an object and pFoo for a pointer.

Patch-set: 1
Reviewer: Gerrit User 1000135 <1000135@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2
Label: SUBM=+1