Update patch set 1
Patch Set 1:

You are very brave for stepping into this can of worms :-)

In lcl_ConvertTimeZone() in sax/source/tools/converter.cxx, you are attempting to convert times between timezones.
I don't think this can be done safely without access to a DST (Daylight Savings Time) database.

You need to know what DST offset was in operation at that specific source date, in order to be able to convert it to a UTC time, 
and then you need to know the DST offset for the destination date in order to correctly switch to it.

See here for a good discussion:
http://stackoverflow.com/questions/2532729/daylight-saving-time-and-timezone-best-practices

The specific recommendation from that stackoverflow article that we should probably follow is that,
if we are going to store a TZ offset with our Date/Time objects, we should also be storing a DST offset.

Also, if we're going to create new commonly used types/structs, perhaps we can use shorter names, like DateTimeTZ?

Patch-set: 1
Reviewer: Gerrit User 1000083 <1000083@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0
Label: Code-Review=0