Update patch set 1
Patch Set 1:

actually i had a look at "ECMA-376 Part 1 3rd edition"
(is this the relevant reference?) and it has:

<quote>
17.18.9
ST_DateTime (Standard Date and Time Storage Format)

This simple type's contents are a restriction of the W3C XML Schema dateTime datatype.
</quote>

which is apparently for the word processing part;
then there is this about spreadsheets:

<quote>
18.17.4
Dates and Times
Each unique instant in SpreadsheetML time is stored as an ISO 8601-formatted string, which is made up of a
date component, a time component, and a timezone component.

L.2.16.9.1
Date Conversion for Serial Values
All date values stored in cells within a SpreadsheetML file are stored in the ISO 8601 format.
</quote>

yay - a standard with 2 different normative references for date formats!

however can't find any reference to http://www.w3.org/TR/NOTE-datetime  in there - where did the idea come from that that was relevant?

Patch-set: 1