Update patch set 2
Patch Set 2:

I don't think it makes sense for the DateTimeWithTimezone struct's DateTimeInTZ member to have the possibility of having different IsTimeZoned values.

What about:

1) DateTimeBase (or RawDateTime?) -> a copy of the current DateTime: NanoSeconds, Seconds, Minutes, Hours, Day, Month, Year

2) DateTime: derives from DateTimeBase, adds IsTimeZoned bool member.

3) DateTimeWithTimezone has a "DateTimeBase DateTimeInTZ;" member and a "short TimeZone;" member.



I've thought long and hard about the "store in UTC" vs "store in localtime" issue.

IMO, "store in UTC" is preferable:

1) It aligns our API with what (AFAIK) most APIs do.

2) Since our "store in localtime" is really "store in what was the local timezone at the moment that LibreOffice started", and not "store in what the OS tells you is the local timezone right now", we need to tell our users (programmers/scripters using our API) to use _LibreOffice's_ notion of localtime (that is, offset to UTC that defines localtime), rather than the system one's (and provide a UNO service that provides the offset). I'm rather worried about creeping bugs in the whole LibreOffice/extension ecosystem where scripts / extensions will blindly copy a UNO DateTime into a system structure (such as e.g. in Python a time.struct_time) and pass it to a library expecting local time (such as Python's time.mktime), and get (sometimes) wrong results.

This is a concern "only" in long-running processes (or more precisely, processes that keep running across a DST change, the computer travelling, etc), but IMO there are several legitimate scenarios where this happens:

 - People that (like me) leave their computer on in the night with stuff left for the next day still open

 - LibreOffice quickstarter on Windows, probably?

 - A LibreOffice process running on a server (with --headless), and other programs / scripts connect to it through URP

 - laptops & pocket computers (e.g. smartphones, tablets, ...) put to hibernation before travel by plane and woken thereafter

3) More generally, "store in startup locatime" forces our users to deal with a *third* timezone; they already have to deal with UTC and with localtime-as-understood-by-system, now we introduce them a different timezone that is nearly, but not quite the same as the latter, namely localtime-as-understood-by-UNO.

4) "store in startup localtime" makes (serialisation of) our UNO DateTime struct NOT SUITABLE for interchange, EVEN ON THE SAME MACHINE.


What problem were you trying to solve by advocating "store in localtime"? From your email:

1) (...) VCL does not have a DateTime widget, but a DateField and a TimeField (...) cannot transparently convert from UTC to local timezone, because they would need both date+time to do that.

Well, yes. As I understand things, these widgets only deal (and until they are changed, *can* *only* deal) with Time values with IsTimeZoned==false. If some other code reads their Value property, it is then that code's responsibility to interpret that as in a specific timezone.

One could argue that since legacy code will generally blissfully ignore the IsTimeZoned member, it gives better backwards compatibility. This assumes that older OO.org/LibreOffice usually served local time in css.util.(Date)Time, which is not obvious to me. It seems to be the case in *some* places, but not in others, so we need a general cleanup anyway. OOXMLDocPropHandler::GetDateTimeFromW3CDTF is an example of a codepath that puts UTC in a UNO DateTime.



Any other reason for "store in startup-time localtime"?

Patch-set: 2
Label: Verified=0
Label: Code-Review=0