gbuild: add an intermediary 'almost' target to Module
This patch is a preparation work for a later patch. that introduce
automatic creation and reuse of binary package.
In order to be able to run a task after every target of a module
have been completed, but without the need to enumerate them all
to that post-module task, an extra intermediary 'target' is introduced
the 'almost' target.
every registered target of a module are depeendencies of 'almost'
and the old done-target for the module now has almsot as dependency

It then become possible to declare a target that is a dependencies of
'done' and has 'almost' as dependencies to guaranty that
this target will be run after the module'w work is completed.

Change-Id: Ifa2bc66e3a5554bf5e85785e1582c57fd62ada5c