Update patch set 1
Patch Set 1: This need some tweaks before it is merged

I have been running some more tests overnight with this patch applied and I do see some memory corruption. Not sure if it comes from this patch, but better be safe and not push it for now.

Patch-set: 1
Label: Code-Review=-1