Update patch set 1
Patch Set 1: This need some tweaks before it is merged

I am not convinced this is necessary. The saving can only be usedfor external projects that are built as ExternalProjects (on all platforms), so all changes in the UnpackedTarball's dir are done by the ExternalProject's targets. Therefore it should suffice to introduce a new target, say gb_UnpackedTarball_save_target, that would depend on all targets registered for the ExternalProject and the ExternalProject's main target would depend on it (this can be easily done in gb_ExternalProject_register_target).

Patch-set: 1
Reviewer: Gerrit User 1000021 <1000021@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0