Use QR decomposition instead of gauss elimination for polynomial reg.
Gauss elimination is not a stable algorithm so use QR decomposition
(with other methods). Also use horner's method to evaluate a polynomial
which is the prefered and more stable method of polynomial evaluation.
Still this is not quite enough as you still have to multiply very large
number with very small and you lose percision on high-degree polynomials
or big X values. Some methods evaluate the polynomial in barycentric form
but how to get the polynomial into the barycentric form is now the
question.

Change-Id: If0d93bc1f08253f015e814e187b5a2cb7f78ec49