Update patch set 3
Patch Set 3: Verified; Looks good to me, approved

Looks good, but the contents of the listboxes got lost. The "stringlists" of old ListBoxes in .src files can be described in .ui as GtkComboBoxText and fill the list into the "items" of that. (There is another more complicated way to do it with GtkComboBox by adding a ComboBox Model to them, but we only need to do that if we need to, e.g. associated data with an entry)

Patch-set: 3
Reviewer: Gerrit User 1000135 <1000135@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2
Label: SUBM=+1