Update patch set 3
Patch Set 3:

> One could argue that since legacy code will generally blissfully
 > ignore the IsTimeZoned member, it gives better backwards
 > compatibility. This assumes that older OO.org/LibreOffice usually
 > served local time in css.util.(Date)Time, which is not obvious to me.
 > It seems to be the case in *some* places, but not in others, so we
 > need a general cleanup anyway.
 > OOXMLDocPropHandler::GetDateTimeFromW3CDTF is an example of a codepath
 > that puts UTC in a UNO DateTime.

I'd regard that as an exception to the general handling, it is not only
*some* places but *most* places that assume local time.

 > Any other reason for "store in startup-time localtime"?

Document persistence and API compatibility. But, I'd not even call that
"startup-time localtime" because calculations of offsets may always be
influenced by the *current* local timezone and DST. I'd rather call it
"wallclock time" because Time/DateTime fields and cell values so far
contain a time not bound to any region. Setting/getting those via UNO
API always assumed that. Reading/writing ODF assumes it. Passing an UTC
time with IsTimeZoned=true will make existing unchanged code assume the
UTC time to be wallclock time instead.

To me the IsTimeZoned member doesn't make much sense other than if it is
set to false. The application does not know the actual offset, no matter
if we define a) true==local_time_with_timezone or b)
true==UTC_with_timezone, it will always use the current offset
using/displaying different values in different timezones for b) or
calculating different UTC for a). For only this reason b) would be
better because the actual used value would be UTC converted to current
local time. But then, if that is wanted, why not use
DateTimeWithTimeZone and TimeZone=0 ?

Btw, I'm tempted to rename the structs and members and .idl to not use
TimeZone but Timezone instead. I find it a tad confusing visually to see
DateTimeWithTimeZone (a date and time with time and zone ...) and would
prefer DateTimeWithTimezone, but maybe that's just me.. DateTimeWithTZ
would also do ;)

Patch-set: 3
Reviewer: Gerrit User 1000023 <1000023@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0
Label: Code-Review=0