Update patch set 5
Patch Set 5: This need some tweaks before it is merged

Marking -1 so that it is not pushed yet :-)

Prashant - it is not finished yet.  Still I can see

    mpFtPosX(new FixedText(this, SVX_RES(FT_SBSHAPE_HORIZONTAL))),
    mpMtrPosX(new MetricField(this, SVX_RES(MF_SBSHAPE_HORIZONTAL))),
    mpFtPosY(new FixedText(this, SVX_RES(FT_SBSHAPE_VERTICAL))),
    mpMtrPosY(new MetricField(this, SVX_RES(MF_SBSHAPE_VERTICAL))),

This must be taken from the .ui; this way you are taking it from the .src (see the SVX_RES usage).

This is one of the panels where you can actually remove the entire .src, there is nothing left there that should not be done by the .ui, so please convert even these widgets.  Thank you!

Patch-set: 5
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0
Label: Code-Review=-1