Modify InsertTable and TableToText methods
The mentioned methods have been modified in order to use the
table style provided as parameter, if present, and the default
frame format if not.

Furthermore, the style is just assigned to the table, rather than
applying the properties immediately. Currently, the style is not
applied, but it will be at render time once the neccesary modifications
are made.

Change-Id: I982ad0539da27e87f3d7443e6a9756503d83be27