gbuild: update included dep-targets as side effect of targets
Make builds in 2 phases: first it ensures that all targets that are
included by the makefiles (i.e. the .d files) are up to date, then
it restarts and builds everything else.

It would be possible to update the .d files in the same build by simply
having dependencies from the dep-targets to the corresponding real targets,
but that has undesirable effects (8b5a984d45005d3df1c89eae897d6e04612625d8).

For a successful build, the .d files can actually be updated as a side
effect of other rules in such a way that they will be up-to-date on the
next make invocation, thus the restart can be avoided.
This optimization will not always work in case of build failures.

Change-Id: I753039ed33f5a84104f35990769eddc83e1d0d7e