Update patch set 1
Patch Set 1: Verified; Looks good to me, but someone else must approve

It basically puts there the orinal code. It just removes the "if" condition because there was the same for cycle in "if" and "else" part. It works fine for me with the patch.

Two more approvals are needed for 4-1-0 branch.

Patch-set: 1
Reviewer: Gerrit User 1000053 <1000053@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+1