Update patch set 1
Patch Set 1: This need some tweaks before it is merged

I would like to hear opinion from from the design team (Stefan) before we do this change in 4.1.

I am not sure of the typical workflow with the autoplay presentations. I think about speakers that do not want to be interrupted by the LO GUI and just want to see the slides. They might be against this change.

A better solution would be to fix opening the file:

   + autoplay presentation when it is opened from the 
     command line

   + edit the presentation when it is opened from LO editor 
     (Impress)

I think that people are used to work this way, see http://www.dailygyan.com/2008/04/how-to-edit-powerpoint-show-pps-or-ppsx.html
     editor

Patch-set: 1
Reviewer: Gerrit User 1000053 <1000053@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0
Label: Code-Review=-1