Update patch set 1
Patch Set 1:

My undersdtanding is that it reverts the code back to the state in 2001, see http://cgit.freedesktop.org/libreoffice/core/commit/?id=1903d481790c799a110cda43a436e472ab554aad

There was:

--- cut ---
- for( sal_Int32 nIdx = 0; nIdx < nCount; nIdx++ )
- pState[nIdx] = getPropertyState( pNames[nIdx] );
+ if( mpImpl->mpMaster )
+ {
+ for( sal_Int32 nIdx = 0; nIdx < nCount; nIdx++ )
+ pState[nIdx] = getPropertyState( pNames[nIdx] );
+
+ }
+ else
+ {
+ for( sal_Int32 nIdx = 0; nIdx < nCount; nIdx++ )
+ pState[nIdx] = getPropertyState( pNames[nIdx] );
+ }
--- cut ---

This change added the if condition but still the same for-cycle was called in both ways.

If I am not blind, David did a long cycle back to the pre-2001 code.

Patch-set: 1