Update patch set 1
Patch Set 1:

I still don't get it.. merging the logic from SvxShape::getPropertyState() and that now changed loop it results in

    for( sal_Int32 nIdx = 0; nIdx < nCount; nIdx++ )
    {
        if( mpImpl->mpMaster )
        {
            pState[nIdx] = mpImpl->mpMaster->getPropertyState( pNames[nIdx] );
        }
        else
        {
            pState[nIdx] = _getPropertyState( pNames[nIdx] );
        }
    }

which to me is not different from the current code except that it determines the condition inside the loop instead of outside. Note the leading underscore for the !mpImpl->mpMaster case in both versions, it is not the same method called.

Patch-set: 1