Update patch set 1
Patch Set 1: Fails; This need some tweaks before it is merged

The filtering works with the patch but I have lost the pivot table when refreshing.

Steps to reproduce:

1. Open the original test document
  "Pivot table with data groups.ods" from
  https://bugs.freedesktop.org/show_bug.cgi?id=63998

2. Press right mouse button in the cell [A5] and select 
   "Refresh"

Result: The pivot table is broken. The data are missing. It shows only the titles, so it has 3 rows and 2 columns:

      Sum - TimeSpent	Billable
      Years	        Total Result
      Total Result


Result without the patch: It shows large table with all 
     non-filtered data.

Expoected result: It should show the original table with 8 
      rows and 7 columns


Note that I do not see this problem when I change the filter in the cell [B3] before doing the refresh. I guess that a pointer is not-initialized correctly during the start or so.

Anyway, the new problem is a kind of data loss => we should not push this patch for 4-1-0 in this state.

Note that the problem has already been in 4.0. I think that the proper fix could wait for 4.1.1 bugfix release.

Patch-set: 1