Update patch set 1
Patch Set 1:

I am not sure. It might be an older bug but I do not see it without this patch. So, if this patch opens a door to a pivottable malformation bug, I would prefer to fix the mailformation before we open the door.

The patch is great step forward, definitely. I am just not brave enough to push it into the last 4.1.0 RC before we know more about the new problem. I am sorry for my caution.

Patch-set: 1