Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

Right click on e.g. a rectangle in draw and it works for me.

for the previews, I think balancing the dialog to have the preview take the same width as the controls and half the height looks about right. I'll try that in a follow up commit.

Also the a11y "labeled by" and "label for" stuff and the "accessibility name" can all be done in the .ui as well, see the follow commit for an example and the "wheelchair" accessibility tab in glade

Patch-set: 1
Reviewer: Gerrit User 1000135 <1000135@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2
Label: SUBM=+1