Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I'm confused.. I thought I saw this only with the patch, but I just checked with my freshly built 4-1-0 and it happens also without and it is the same in 4.0.4. Important is that one does not select a Project Name filter first (maybe I failed at that), if one did all entries are shown also after Refresh. So this behavior is not related to this patch, I'm giving my +1 again.

Patch-set: 1
Label: Code-Review=+1