Update patch set 3
Patch Set 3: This need some tweaks before it is merged

This still isn't quite right for merging. But now it appears for me without a crash, though a few bits of functionality aren't working yet.

I wonder if you are not building with --enable-dbgutil. With --enable-dbgutil you get error messages like "element foo should be a FixedText" and hints like that.

I wonder if instead of moving the "optionshelper" code into this cxx if it would be better to duplicate it as a "newoptionshelper" because it seems the other filter dialog uses the old one so when that dialog is converted over as well it might be able to use the new helper

Patch-set: 3
Label: Verified=0
Label: Code-Review=-1