Update patch set 3
Patch Set 3:

Seeing noone reviews the real changes in this patch, maybe this should split up: the editeng and svx changes in one patch and filter parts in an other?

I have no idea about the editeng/svx parts, but if they are in, I can do the review of the filter parts.

Patch-set: 3
Label: Verified=0
Label: Code-Review=0