Update patch set 1
Patch Set 1:

That can only happen, if in dmapper, PopFootOrEndnote() is called without PushFootOrEndnote() first. This is possible, if there is a bug in the tokenizers, but ATM the DOCX/RTF tokenizers ensure these are called in pairs, with whatever user input. I'll add a check for it, though, just to be sure.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0
Label: Code-Review=0