WaE: may be used uninitialized
For some reason, GCC 4.7 targeting X86 in the Android NDK 8e complained about
the fields in the "btrd" struct in the union in binomial_distribution
(boost/random/binomial_distribution.hpp) maybe being uninitialized in a call
to its copy constructor. So make sure these fields are initialized in all
cases.

Change-Id: Iebae5e130cd6ca6ef5c476c6b0366d57c0303c88