Update patch set 1
Patch Set 1: This need some tweaks before it is merged

I think, that this should not be done.
(1) If you insert a background (www-back) from the Gallery to a Writer/Web document, it is always linked. Removing the source might break old documents.
(2) If you insert a background as page background from the Gallery to a Writer document, it is linked as default. Removing the source might break old documents.
(3) If you will make a list with graphic bullets, it uses the bullets from the Gallery. Users might have checked the option to link them. Removing the source might break old documents.
(4) Currently the predefined list style sets use the bullets from the Gallery. Unless the list styles are changed the bullets in the Gallery should not be removed.
(5) Removing sounds make no sense at all. There are no new ones.
(6) The pictures "flower" and "apple" should remain, because it is very handy to have common pictures, when you write documentations.

Patch-set: 1
Reviewer: Gerrit User 1000447 <1000447@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1