Update patch set 2
Patch Set 2: Verified; Looks good to me, approved

The thing to do is to make the label always "Shown" so the height is taken into account at the start, but leave it blank and just set its text if there is an error.

Couple of small nits around spacing, i.e. we try and stick to multiples of 6 pixels spacing (or 3 in a pinch). And the title of the dialog was missing.

I'll fix those up in a follow up commit

Patch-set: 2
Reviewer: Gerrit User 1000135 <1000135@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2
Label: SUBM=+1