Fix table styles not loading
Some unused properties were no longer loaded and that messed
up the styles, breaking the first and not loading the others.

Information about font is not loaded properly and selecting one
of the styles for a table causes an error.

Things will most likely be changed again since the targeted ideal
is to create the styles with the document.

Change-Id: I17a2c2c2f4eff36265489f725a83a4317b27339e