Update patch set 1
Patch Set 1:

There is an updated patch on bug 66279 with a cleaner setup (using a separate starmath/source/operatorDictionary.cxx file to store the dictionary rather than an #include). I can update this patch if necessary. However, integrating the generation of the dictionary in the build system may be more work, I'm not very familiar with it and don't have much time to look into that so it's unlikely to happen soon.

Patch-set: 1
Reviewer: Gerrit User 1000449 <1000449@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0
Label: Code-Review=0