Resolves: fdo#67113 split insert autotext and field controls
autotext is a modal dialog and insert field is a modeless dialog so I don't
want autotext to show a toggle in insert->autotext

At the same time we want the autotext entry in the insert toolbar to have a
dropdown and we appear to want to be able to close the insert field dialog from
the toolbar.

So split the combined handler class into two and remove the other if/else
branch in each with the insert field a SfxBoolItem and the insert autotext an
SfxVoidItem.

Change-Id: I3d59052ee70a00d63f7706166290b00ad9ae22c0
(cherry picked from commit bb98778dbf761a0c88c96117add00a66e5cc6c95)