Added a for loop to "CalcAutoLayoutRectangles" to change logic.
Changed the logic of switch case in "CalcAutoLayoutRectangles".This for loop basically traverse through the "layoutlist.xml" and
find the appropriate node to a particular "layout" and then fills in the property values to all of it's presobj in the layout.
Currently,I have not deleted the switch case,but this 'for loop' aims to replace it.

Change-Id: I114614292934af179f9380277021aea350e3b35d