Update patch set 1
Patch Set 1: This need some tweaks before it is merged

Work in progress

There is a glitch in rendering the shadows in the preview window. The rectangles (object and shadow) are not centered in the window.

Patch-set: 1
Reviewer: Gerrit User 1000059 <1000059@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0
Label: Code-Review=-1