fdo#66006 Convert to PNG for JPEG that can't be losslessly rotated
If a JPEG image has a width or height which is not a multiple of 16
(width/heigh % 16 != 0) then it can't be completelly lossleessly
rotated (the last uncomplete MCUs can't be rotated) and the image
is cropped on the lower size. To prevent this such JPEG images
must be rotated to PNG to prevent loss. Afterwards image can still
be compressed with "Compress Graphic Dialog".

Change-Id: Ie2803512a93bba55573bf66c547f132f37cba711
Reviewed-on: https://gerrit.libreoffice.org/5159
Reviewed-by: Luboš Luňák <l.lunak@suse.cz>
Tested-by: Luboš Luňák <l.lunak@suse.cz>