Update patch set 2
Patch Set 2:

The original idea with the tab was that there might be more font-related settings (MSO does have them). Anyway, I don't see that as a blocker for this patch, as it's not a big difference if there are 3 or 4 such tabs, and whoever can later convert 3 can easily convert 4 as well.

Patch-set: 2
Reviewer: Gerrit User 1000092 <1000092@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0