fdo#66715: fontconfig: try harder to ignore duplicate fonts
The thin space not being displayed correctly is caused by using the
wrong font, namely
/usr/share/fonts/liberation/LiberationSerif-Regular.ttf,
which (on Fedora 18) is version 1 and does not contain u2006 etc.
glyphs, whereas the LiberationSerif-Regular.ttf bundled with LO
is version 2 and does contain these.

There is already isPreviouslyDuplicateOrObsoleted() function to ignore
older fonts but it does not work for this case because:

1) Only the previous element was looked at, but there may be several
   fonts with different weight/slant that need to be checked.

2) The LiberationSerif-Regular.ttf differ in the "lang" entry.

Change-Id: I2f9e8d50a1f8155b65f8f07c9259dd988c32992a