Fix missing font attributes in AutoFormat window
The attribute range associated with the SwTableBoxFmt did not
include the necessary ids for character attributes, and as a
result, these were not stored and not used correctly.

The solution should not have any side-effects, but, to be sure
there should be further testing of other features where this
range came into use.

Change-Id: I8cb1ac881b2746ae3ec7958a958167de6f140a7f