fdo#67430 Keep original order of entry positions, not alphabetically sorted
...turns out ca0600f0c9022d631317423ab5a59493b41906ab "multiple unsequenced
modifications to 'nPos' [-Werror,-Wunsequenced]" was wrong in assuming that the
order of getPropertyPos had no meaning, as it determines the order in which
form property dialogs display items.

So, reintroduced a fixed order, but without resorting to undefined behavior due
to repeated uses of ++ with no sequence points in between.  This is rather
sluggish, but gets the job done.

(cherry picked from commit a6fe1fde409ff1accdea49ff9de69658de1e6f5f)
Conflicts:
	extensions/source/propctrlr/formmetadata.cxx

Change-Id: If9c5f4130fe6523eebd3a13e5a685fc9c0f53fc0
Reviewed-on: https://gerrit.libreoffice.org/5215
Reviewed-by: Fridrich Strba <fridrich@documentfoundation.org>
Tested-by: Fridrich Strba <fridrich@documentfoundation.org>