Update patch set 1
Patch Set 1:

Hmm, I don't notice that;)

The difference I see in a LO (current master) with this patch applied compared to one without is that when entering text into Calc cells, in the patched LO there is no "jumping". There still is jumping in "comment boxes". Text caret positioning seems correct to me in all cases. (This is with just Latin script, and the default fonts LO uses.)

So based on the above tests, I would tend to actually want to approve this change;)

Patch-set: 1