fdo#67621: Delay populating the ranges until the dialog is shown.
Otherwise the table widget won't have any idea which rows are displayed,
and end up not populating the ranges.

This is caused by the switch to the .ui dialog, which delays calculation
of the dialog size until late. But the code that populates the ranges was
execuated when the widget was instantiated, at which time the widget didn't
have its size assigned.

Change-Id: I22943b41b21b58cc67e872dff13ad3c25eee8438