Register SwFmts to correct parents for each element of table
Insert table, text to table and set autoformat methods now set the
parent of each SwFmt to the correct element in the table style.
For more complex tables, i.e. boxes contain other lines and so on,
the parents for these are set to the same as the line/box in which
they are found.

Change-Id: I24bc35323414bc8bdceda610da7ba53da043f09b