Fix accessibility of text attributes on OS X
The applyAttributesFrom:... method takes range parameter that specifies
a range in the 'string' parameter passed to that method, not in the whole
string of the UI element as retrievd in createAttributedStringForElement:...
(the former string is a substring of the latter).

In other words, the 'range' parameter in applyAttributesFrom: is relative
to the string passed to that method, not to the whole string of the UI
element.

This enables proper reading of text attributes in Writer - when moving
cursor through text, VoiceOver now announces e.g. "bold" or "plain" to
indicate changes in text formatting.

Change-Id: I21a633af0bf426759f639435581fcd3bfeafa598