Revert "fdo#54320: VoiceOver does not follow keyboard focus"
The code was correct (and docs confusing). The code did not work because
67410 was not fixed at that moment. With 67410 fixed, the code works
as it was originally.

Also in the meantime I discovered a comment in the NSAccessibility.h header
telling that NSAccessibilityPostNotification should be passed the newly focused
element and that the function itself will make sure to post the notification
on the NSApp element instead.

This reverts commit e9257fa0e642de2f1c79f2c80387334010812dab.

Change-Id: I7f76ebde963cf57879273d9848ff195877759524