Update patch set 4
Patch Set 4:

Wow, it seems working :) The only thing that bothers me is the close button doesn't use any underline.
The .hrc should be deleted, but when I set the remove in the git, it always wants to rename the optdict.hrc to optdict.src, is it okay?

Patch-set: 4
Reviewer: Gerrit User 1000463 <1000463@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=0